+++ src/scannerwindow.h
@@ -36,6 +36,7 @@ class QSplitter;
 class QStringListModel;
 class QTableView;
 class QTextEdit;
+class QTextBrowser;
 class QTimer;
 class QAction;
 class QCloseEvent;
@@ -201,6 +202,10 @@ private:
                                              QString *error);
     void updateProbeSummary();
     void setDetailsPaneVisible(bool visible);
+    int defaultDetailsPaneHeight() const;
+    void applyDetailsPaneHeight();
+    void configureExternalLinks(QTextBrowser *browser);
+    void openExternalLink(const QUrl &url);
     void rebuildMainToolbar();
     void applyToolbarDisplayMode();
     bool openPreferredTerminal(const QStringList &args = {}, QString *error = nullptr) const;
@@ -238,6 +243,7 @@ private:
     QLineEdit *searchInput_ = nullptr;
     QSplitter *resultsSplitter_ = nullptr;
     QTextEdit *detailsPane_ = nullptr;
+    int detailsPaneHeight_ = 0;
 
     QLabel *statusTextLabel_ = nullptr;
     QProgressBar *statusProgressBar_ = nullptr;